permissions draft

+++ users/views.py
@@ -2,12 +2,10 @@ from .models import User
 from rest_framework import viewsets
 from .serializers import UserSerializer
 from middlewares.auth import FirebaseAuthentication
-
 from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.decorators import authentication_classes, permission_classes
-
-from helpers.permissions import IsOwner, IsAdmin, IsTest
+from helpers.permissions import IsOwner, IsAdmin, IsOwnerOrAdmin
 
 class UserDetail(viewsets.ViewSet):
 
@@ -17,11 +15,11 @@ class UserDetail(viewsets.ViewSet):
         """
         self.permission_classes = []
         if self.action == 'retrieve_all': self.permission_classes = [IsAdmin]
-        if self.action == 'retrieve': self.permission_classes = [IsAdmin|IsOwner]
+        if self.action == 'retrieve': self.permission_classes = [IsOwnerOrAdmin]
         if self.action == 'create': self.permission_classes = []
-        if self.action == 'update': self.permission_classes = [IsAdmin]
-        if self.action == 'partial_update': self.permission_classes = [IsAdmin]
-        if self.action == 'destroy': self.permission_classes = [IsAdmin]
+        if self.action == 'update': self.permission_classes = [IsOwnerOrAdmin]
+        if self.action == 'partial_update': self.permission_classes = [IsOwnerOrAdmin]
+        if self.action == 'destroy': self.permission_classes = [IsOwnerOrAdmin]
         return [permission() for permission in self.permission_classes]
 
     def get_authenticators(self):
@@ -53,6 +51,7 @@ class UserDetail(viewsets.ViewSet):
     def update(self, request, pk=None):
         try:
             instance = User.objects.get(pk=pk)
+            self.check_object_permissions(self.request, instance)
         except User.DoesNotExist:
             return Response('User Does not exist', status=status.HTTP_404_NOT_FOUND)
         serializer = UserSerializer(instance, data=request.data)
@@ -65,6 +64,7 @@ class UserDetail(viewsets.ViewSet):
     def partial_update(self, request, pk=None):
         try:
             instance = User.objects.get(pk=pk)
+            self.check_object_permissions(self.request, instance)
         except User.DoesNotExist:
             return Response('User Does not exist', status=status.HTTP_404_NOT_FOUND)
         serializer = UserSerializer(instance, data=request.data, partial=True)
@@ -77,6 +77,7 @@ class UserDetail(viewsets.ViewSet):
     def destroy(self, request, pk=None):
         try:
             instance = User.objects.get(pk=pk)
+            self.check_object_permissions(self.request, instance)
             instance.delete()
         except User.DoesNotExist:
             return Response('User Does not exist', status=status.HTTP_404_NOT_FOUND)
